backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
gensim==4.3.2
umap-learn==0.5.4
scikit-learn==1.3.0
numpy==1.25.2
pydantic==2.3.0
cachetools==5.3.1
